+++ src/api/alchemy-provider.ts
@@ -10,6 +10,8 @@ import {
 } from '@ethersproject/providers';
 import { ConnectionInfo, fetchJson } from '@ethersproject/web';
 
+import { JsonRpcRequest, JsonRpcResponse } from '../internal/internal-types';
+import { RequestBatcher } from '../internal/request-batcher';
 import { Network } from '../types/types';
 import {
   CustomNetworks,
@@ -38,6 +40,13 @@ export class AlchemyProvider
 {
   readonly apiKey: string;
   readonly maxRetries: number;
+  readonly batchRequests: boolean;
+
+  /**
+   * VISIBLE ONLY FOR TESTING
+   *@internal
+   */
+  readonly batcher: RequestBatcher;
 
   /** @internal */
   constructor(config: AlchemyConfig) {
@@ -68,6 +77,17 @@ export class AlchemyProvider
 
     this.apiKey = config.apiKey;
     this.maxRetries = config.maxRetries;
+    this.batchRequests = config.batchRequests;
+
+    // TODO: support individual headers when calling batch
+    const batcherConnection = { ...this.connection };
+    batcherConnection.headers!['Alchemy-Ethers-Sdk-Method'] = 'batchSend';
+    const sendBatchFn = (
+      requests: JsonRpcRequest[]
+    ): Promise<JsonRpcResponse[]> => {
+      return fetchJson(batcherConnection, JSON.stringify(requests));
+    };
+    this.batcher = new RequestBatcher(sendBatchFn);
   }
 
   /**
@@ -218,13 +238,27 @@ export class AlchemyProvider
    *
    * @internal
    */
-  _send(method: string, params: Array<any>, methodName: string): Promise<any> {
+  _send(
+    method: string,
+    params: Array<any>,
+    methodName: string,
+    forceBatch = false
+  ): Promise<any> {
     const request = {
       method,
       params,
       id: this._nextId++,
       jsonrpc: '2.0'
     };
+
+    // START MODIFIED CODE
+    const connection = { ...this.connection };
+    connection.headers!['Alchemy-Ethers-Sdk-Method'] = methodName;
+
+    if (this.batchRequests || forceBatch) {
+      return this.batcher.enqueueRequest(request as JsonRpcRequest);
+    }
+    // END MODIFIED CODE
 
     this.emit('debug', {
       action: 'request',
@@ -238,11 +272,6 @@ export class AlchemyProvider
     if (cache && this._cache[method]) {
       return this._cache[method];
     }
-
-    // START MODIFIED CODE
-    const connection = { ...this.connection };
-    connection.headers!['Alchemy-Ethers-Sdk-Method'] = methodName;
-    // END MODIFIED CODE
 
     const result = fetchJson(
       this.connection,