+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Minor Changes
 
 - Added support for ENS resolution to the `NotifyNamespace` and `WebsocketNamespace` methods.
+- Included new export for the `Interface` class from ethers to simplify encoding ABI when using `eth_call`.
 
 ## 2.2.1